2023/CSAW-Quals/misc/AdminDiscordBot/requirements.txt
discord.py==1.5.0
python-dotenv==1.0.0
